% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Looney and Jones' Corrected Z-test.R
\name{corrected.z.test}
\alias{corrected.z.test}
\title{Looney and Jones' Corrected Z-test}
\usage{
corrected.z.test(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values with some missing value(NA).}

\item{y}{a (non-empty) numeric vector of data values with some missing value(NA).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
Results of test including p-value will be printed.
}
\description{
Performs Looney and Jones's Corrected Z-test on a partially matched sample in a form of R vector.
Depending on the number of missing values in the two input vectors, the test can become Two Sample Z-test or Paired Sample Z-test.
}
\examples{
# Generating Toy Examples
set.seed(123)
x <- rnorm(20)
x[sample(1:20, 3)] <- NA # Deliverately generating some missing values
y <- (rnorm(20) + 1)/3
y[sample(which(!is.na(x)), 4)]
corrected.z.test(x, y)
corrected.z.test(x, y, alternative = "greater")
corrected.z.test(x, y, alternative = "less")

}
