% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin and Stiver's MLE Based Test Under
%   Heteroscedasticity.R
\name{lin.stiver.test}
\alias{lin.stiver.test}
\title{Lin and Stiver's MLE Based Test Under Heteroscedasticity}
\usage{
lin.stiver.test(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values with some missing value(NA).}

\item{y}{a (non-empty) numeric vector of data values with some missing value(NA).}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
Results of test including p-value will be printed.
}
\description{
Performs Lin and Stiver's MLE Based Test Under Heteroscedasticity on a partially matched sample in a form of R vector.
This test is under the assumption of unequal variance of the two input vectors. By F-test, if the two vectors are considered to have the same variance, a warning message will be shown after the result is printed.
}
\examples{
# Generating Toy Examples
set.seed(123)
x <- rnorm(20)
x[sample(1:20, 3)] <- NA # Deliverately generating some missing values
y <- (rnorm(20) + 1)/3
y[sample(which(!is.na(x)), 4)] <- NA
lin.stiver.test(x, y)
lin.stiver.test(x, y, alternative = "greater")
lin.stiver.test(x, y, alternative = "less")

}
